requirements.txt
dill
opencv
ipython
numpy
pandas
matplotlib
ipywidgets
setuptools
pillow